When allow retry, login as admin should be always done by default
